# perf(modeling): faster toOutlines