--- a/packages/modeling/src/operations/extrusions/slice/calculatePlane.js
+++ b/packages/modeling/src/operations/extrusions/slice/calculatePlane.js
@@ -16,7 +16,7 @@ const calculatePlane = (slice) => {
   if (edges.length < 3) throw new Error('slices must have 3 or more edges to calculate a plane')
 
   // find the midpoint of the slice, which will lie on the plane by definition
-  const midpoint = edges.reduce((point, edge) => vec3.add(vec3.create(), point, edge[0]), vec3.create())
+  const midpoint = edges.reduce((point, edge) => vec3.add(point, point, edge[0]), vec3.create())
   vec3.scale(midpoint, midpoint, 1 / edges.length)
 
   // find the farthest edge from the midpoint, which will be on an outside edge