Performance improvements for `toOutlines`. This is used internally by `extrudeLinear`, `extrudeRectangular`, and `offsetGeom2`. Changes include:

- Use `Map` instead of `{}`
- Split up large `toOutlines` function into smaller `toSharedVertices`, `toVertexMap` and `popNextEdge` functions
- Remove unnecessary allocations of `vec3.create()`
- Remove unnecessary allocations in `geom3ToGeometries`

This results in performance improvements when running `toOutlines` on a circle:

1k segments: 2.5ms -> 2.1ms
10k segments: 18ms -> 14ms
50k segments: 92ms -> 72ms
100k segments: 172ms -> 125ms

### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?